Make imports of JNI-defined modules conditional
We import Android stuff only if the (kind-of phony, see below) `android`
module can be imported first. That module is defined in CPython here,
https://github.com/aaaaalbert/sensibility-testbed/blob/0862f5f8b6e5ed01d014ef33fa25bdd8bb0891da/app/src/main/jni/popen.c#L85

This module is "kind of phony" because it doesn't yet suffice to import
it and then get access to all the sensors etc.; in fact, lots of prior
initialization *and then* some more CPython imports are required for
sensors to work. For the namespace modifications at hand, we actually
exploit the side effect of the `android` module only existing in our
Android platform distribution.